feat: トップ画面を作成し、全体のデザインを整えました
