--- frontend/src/pages/hello_mui.tsx
@@ -1,12 +1,0 @@
-import { Button } from '@mui/material'
-import type { NextPage } from 'next'
-
-const HelloMui: NextPage = () => {
-  return (
-    <>
-      <Button>Hello Mui@v5!</Button>
-    </>
-  )
-}
-
-export default HelloMui
